Undo refactor and more closely follow other APIs
