+++ packages/playwright/src/matchers/toMatchText.ts
@@ -15,11 +15,17 @@
  */
 
 
-import { expectTypes } from '../util';
+import { expectTypes, callLogText } from '../util';
+import {
+  printReceivedStringContainExpectedResult,
+  printReceivedStringContainExpectedSubstring
+} from './expect';
+import { EXPECTED_COLOR } from '../common/expectBundle';
 import type { ExpectMatcherState } from '../../types/test';
+import { kNoElementsFoundError, matcherHint } from './matcherHint';
 import type { MatcherResult } from './matcherHint';
 import type { Locator } from 'playwright-core';
-import { textMatcherMessage, toMatchExpectedStringOrPredicateVerification } from './error';
+import { colors } from 'playwright-core/lib/utilsBundle';
 
 export async function toMatchText(
   this: ExpectMatcherState,
@@ -31,7 +37,23 @@ export async function toMatchText(
   options: { timeout?: number, matchSubstring?: boolean } = {},
 ): Promise<MatcherResult<string | RegExp, string>> {
   expectTypes(receiver, [receiverType], matcherName);
-  toMatchExpectedStringOrPredicateVerification(this, matcherName, receiver, receiver, expected);
+
+  const matcherOptions = {
+    isNot: this.isNot,
+    promise: this.promise,
+  };
+
+  if (
+    !(typeof expected === 'string') &&
+    !(expected && typeof expected.test === 'function')
+  ) {
+    // Same format as jest's matcherErrorMessage
+    throw new Error([
+      matcherHint(this, receiver, matcherName, receiver, expected, matcherOptions),
+      `${colors.bold('Matcher error')}: ${EXPECTED_COLOR('expected',)} value must be a string or regular expression`,
+      this.utils.printWithType('Expected', expected, this.utils.printExpected)
+    ].join('\n\n'));
+  }
 
   const timeout = options.timeout ?? this.timeout;
 
@@ -46,24 +68,52 @@ export async function toMatchText(
   }
 
   const stringSubstring = options.matchSubstring ? 'substring' : 'string';
+  const receivedString = received || '';
+  const messagePrefix = matcherHint(this, receiver, matcherName, 'locator', undefined, matcherOptions, timedOut ? timeout : undefined);
+  const notFound = received === kNoElementsFoundError;
+
+  let printedReceived: string | undefined;
+  let printedExpected: string | undefined;
+  let printedDiff: string | undefined;
+  if (pass) {
+    if (typeof expected === 'string') {
+      if (notFound) {
+        printedExpected = `Expected ${stringSubstring}: not ${this.utils.printExpected(expected)}`;
+        printedReceived = `Received: ${received}`;
+      } else {
+        printedExpected = `Expected ${stringSubstring}: not ${this.utils.printExpected(expected)}`;
+        const formattedReceived = printReceivedStringContainExpectedSubstring(receivedString, receivedString.indexOf(expected), expected.length);
+        printedReceived = `Received string: ${formattedReceived}`;
+      }
+    } else {
+      if (notFound) {
+        printedExpected = `Expected pattern: not ${this.utils.printExpected(expected)}`;
+        printedReceived = `Received: ${received}`;
+      } else {
+        printedExpected = `Expected pattern: not ${this.utils.printExpected(expected)}`;
+        const formattedReceived = printReceivedStringContainExpectedResult(receivedString, typeof expected.exec === 'function' ? expected.exec(receivedString) : null);
+        printedReceived = `Received string: ${formattedReceived}`;
+      }
+    }
+  } else {
+    const labelExpected = `Expected ${typeof expected === 'string' ? stringSubstring : 'pattern'}`;
+    if (notFound) {
+      printedExpected = `${labelExpected}: ${this.utils.printExpected(expected)}`;
+      printedReceived = `Received: ${received}`;
+    } else {
+      printedDiff = this.utils.printDiffOrStringify(expected, receivedString, labelExpected, 'Received string', false);
+    }
+  }
+
+  const message = () => {
+    const resultDetails = printedDiff ? printedDiff : printedExpected + '\n' + printedReceived;
+    return messagePrefix + resultDetails + callLogText(log);
+  };
 
   return {
     name: matcherName,
     expected,
-    message: () =>
-      textMatcherMessage(
-          this,
-          matcherName,
-          receiver,
-          'locator',
-          expected,
-          received,
-          log,
-          stringSubstring,
-          pass,
-          !!timedOut,
-          timeout,
-      ),
+    message,
     pass,
     actual: received,
     log,
